Added ability to display fonts on the word clocks display using the letters as dots.
Due to the fact that the Arduino has very limited memory, the fonts are managed as
byte arrays, created by an utility program. The utility program (createfont.c) uses
font definitions in ASCII to generate the byte array for the Arduino Sketch program.

The current implementation is limited to fonts which are 8 or less pixel wide as a
single byte is used to store the pixels of one row.